refactor: 🚚 abstract view and logic
