+++ apps/ui/src/components/pages/application/ApplicationViewer.tsx
@@ -16,24 +16,42 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-import { Route, Routes } from 'react-router';
 
-import PageLayout from '@/components/layouts/PageLayout';
-import DashboardPage from '@/pages/dashboard';
-import HomePage from '@/pages/index';
-import ApplicationIndex from './application';
+import { Col, Row } from 'antd';
+import { Content } from 'antd/es/layout/layout';
 
-function App() {
+import { Application } from '@pcgl-daco/data-model';
+
+import { contentWrapperStyles } from '@/components/layouts/ContentWrapper';
+import { SkeletonLoader } from '@/components/SkeletonLoader';
+
+type ApplicationViewerProps = {
+	isEditMode: boolean;
+	data: Application | undefined;
+	loading: boolean;
+};
+
+function ApplicationViewer({ isEditMode, data, loading }: ApplicationViewerProps) {
 	return (
-		<Routes>
-			<Route element={<PageLayout />}>
-				<Route index element={<HomePage />} />
-				<Route path="dashboard" element={<DashboardPage />} />
-				<Route path="application/:id" element={<ApplicationIndex isEditMode={false} />} />
-				<Route path="application/:id/edit" element={<ApplicationIndex isEditMode={true} />} />
-			</Route>
-		</Routes>
+		<Content>
+			<Row style={{ ...contentWrapperStyles }}>
+				{loading ? (
+					//Loading state.
+					//TODO: Temporary, but we should make this look pretty.
+					<SkeletonLoader />
+				) : (
+					<Col>
+						<p>
+							Mode is: <strong>{isEditMode ? ' Edit Mode' : ' View Mode'}</strong>
+						</p>
+						<h1>PCGL-{data.id}</h1>
+						<h2>Application Created - {data.created_at.toLocaleString('en-CA')}</h2>
+						<p>Not set up yet.</p>
+					</Col>
+				)}
+			</Row>
+		</Content>
 	);
 }
 
-export default App;
+export default ApplicationViewer;
